+++ sign/tbls/tbls.go
@@ -115,7 +115,7 @@ func (s *scheme) VerifyRecovered(public kyber.Point, msg, sig []byte) error {
 // can be verified through the regular BLS verification routine using the
 // shared public key X. The shared public key can be computed by evaluating the
 // public sharing polynomial at index 0.
-func (s *scheme) Recover(public *share.PubPoly, msg []byte, sigs [][]byte, t, n int) ([]byte, error) {
+func (s *scheme) Recover(public *share.PubPoly, msg []byte, sigs [][]byte, t, n uint32) ([]byte, error) {
 	var pubShares []*share.PubShare
 	for _, sig := range sigs {
 		sh := SigShare(sig)
@@ -132,11 +132,11 @@ func (s *scheme) Recover(public *share.PubPoly, msg []byte, sigs [][]byte, t, n 
 			continue
 		}
 		pubShares = append(pubShares, &share.PubShare{I: idx, V: point})
-		if len(pubShares) >= t {
+		if uint32(len(pubShares)) >= t {
 			break
 		}
 	}
-	if len(pubShares) < t {
+	if uint32(len(pubShares)) < t {
 		return nil, errors.New("not enough valid partial signatures")
 	}
 	commit, err := share.RecoverCommit(s.sigGroup, pubShares, t, n)
